+++ app/components/dungeons/PrizeList.tsx
@@ -36,7 +36,7 @@ export class PrizeList extends React.PureComponent<Props> {
                 data-tip={showTooltips}
                 data-for={showTooltips ? `prize-tooltip-${prize.id}` : undefined}
               >
-                <img src={itemImage(lang, prize.id, prize.type)} width={32} height={32} />
+                <img src={itemImage(lang, prize.id, prize.type)} width={32} height={32} alt="" />
                 {prize.name} ×{formatAmount(prize.amount)}
                 {showTooltips && prize.type === ItemType.Relic && enlir.relics[prize.id] && (
                   <RelicTooltip